minor reorganization - attempt to build a privacy wallet library to share code
