+++ utt/CMakeLists.txt
@@ -194,6 +194,7 @@ include(dependencies.cmake)
 add_subdirectory(libutt)
 add_subdirectory(libxassert)
 add_subdirectory(libxutils)
+
 if(BUILD_WALLET_CLI)
     add_subdirectory(wallet-cli) 
 endif()
@@ -203,9 +204,7 @@ endif()
 if(BUILD_WALLET_CLI AND BUILD_PRIVACY_ADMIN_CLI)
     add_subdirectory(e2e)
 endif()
-if(BUILD_PRIVACY_WALLET_SERVICE)
-    add_subdirectory(privacy-wallet-service)
-endif()
+
 # [TODO-UTT] This improved api for libutt could go into its own subproject 
 set(newutt_src 
     libutt/src/api/UTTParams.cpp
@@ -226,6 +225,12 @@ target_link_libraries(utt_api PUBLIC utt)
 target_include_directories(utt_api PUBLIC include)
 
 add_subdirectory(utt-client-api)
+
+add_subdirectory(privacy-wallet-lib)
+
+if(BUILD_PRIVACY_WALLET_SERVICE)
+    add_subdirectory(privacy-wallet-service)
+endif()
 
 add_subdirectory(tests)
