+++ utt/privacy-wallet-service/src/Wallet.cpp
@@ -12,7 +12,7 @@
 // file.
 
 #include "Wallet.hpp"
-#include "FileBasedUserStorage.hpp"
+#include <storage/FileBasedUserStorage.hpp>
 #include <iostream>
 #include <utt-client-api/ClientApi.hpp>
